handle short files better, add progress output
Files are read until the first timestamp is found so start_time
can be set, but some files don't have any data rows so the call
to __next__() on the csv.reader object was failing with
a StopIteration exception. Catch that and throw a FlightLogException
instead.

I tried using cht_ranges.py on 1000 log files, and it was taking a while,
so I added a "n of m" progress output display.

I also added some comments to the top of cht_ranges.py.